% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into.R
\name{insert_into}
\alias{insert_into}
\title{insert_into}
\usage{
insert_into(x, y, where = 1)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame}

\item{where}{position to insert}
}
\value{
A modified data.frame
}
\description{
Used by \code{\link[=add_col]{add_col()}}
}
\details{
insert one data.frame into another in between the column
}
\examples{
df1 <- data.frame(x = 1:2, y = c("a", "b"))
df2 <- data.frame(z = c(TRUE, FALSE))
insert_into(df1, df2, where = 1)
}
\seealso{
\code{\link[=add_col]{add_col()}}

Other xyz: \code{\link{add_col}}
}
\concept{xyz}
